+++ R/gComp.R
@@ -13,7 +13,7 @@
 #'   confidence intervals. The \code{clusterID} will be used as the level for
 #'   resampling in the bootstrap procedure.
 #' @param parallel (Optional) Default "no." The type of parallel operation to be used. Available 
-#'   options (besides the default of no parallel processing) include multicore" (not available 
+#'   options (besides the default of no parallel processing) include "multicore" (not available 
 #'   for Windows) or "snow." This argument is passed directly to \code{\link[boot]{boot}}.
 #'   See note below about setting seeds and parallel computing.
 #' @param ncpus (Optional, only used if parallel is set to "multicore" or "snow") Default 1. 
@@ -104,9 +104,19 @@
 #'   the variable. However, variations in the effect are likely small, 
 #'   especially near the mean.
 #'        
+#' @note       
+#'  Interaction terms are not allowed in the model formula. The \code{subgroup} 
+#'   argument affords interaction between the exposure variable and a single 
+#'   covariate (that is forced to categorical if supplied as numeric) to 
+#'   estimate effects of the exposure within subgroups defined by the 
+#'   interacting covariate. To include additional interaction terms with 
+#'   variables other than the exposure, we recommend that users create the 
+#'   interaction term as a cross-product of the two interaction variables in  
+#'   a data cleaning step prior to running the model.
+#'        
 #' @note 
 #'  The documentation for \code{\link[boot]{boot}} includes details about 
-#'  reproducible seeds when using parallel computing.   
+#'   reproducible seeds when using parallel computing.   
 #'  
 #' @export
 #' 
@@ -161,6 +171,8 @@
 #'
 #' @seealso \code{\link{pointEstimate}} \code{\link[boot]{boot}}
 #'   
+#'   
+
 gComp <- function(data, 
                   outcome.type =  c("binary", "count","count_nb", "rate", "rate_nb", "continuous"), 
                   formula = NULL,