+++ R/pointEstimate.R
@@ -53,7 +53,7 @@
 #'   Poisson regression model. Numeric variable should be on the linear scale; 
 #'   function will take natural log before including in the model.
 #' @param rate.multiplier (Optional, only applicable for rate/count outcomes). 
-#'   Default 1 Numeric variable signifying the person-time value to use in 
+#'   Default 1. Numeric variable signifying the person-time value to use in 
 #'   predictions; the offset variable will be set to this when predicting under 
 #'   the counterfactual conditions. This value should be set to the person-time 
 #'   denominator desired for the rate difference measure and must be inputted in 
@@ -100,18 +100,17 @@
 #'   \enumerate{ 
 #'   \item Fit a regression of the outcome on the exposure
 #'   and relevant covariates, using the provided data set. 
-#'   \item Using the
-#'   modelfit in step 1, predict counterfactuals (e.g. calculate predicted
-#'   outcomes for each observation in the data set under each level of the
-#'   treatment/exposure). 
+#'   \item Using the model fit in step 1, predict counterfactuals (e.g. 
+#'   calculate predicted outcomes for each observation in the data set under 
+#'   each level of the treatment/exposure). 
 #'   \item Estimate the marginal difference/ratio of treatment effect by 
 #'   taking the difference or ratio of the average of all observations under 
 #'   the treatment/no treatment regimes. 
 #'   }
 #'
 #'   As counterfactual predictions are generated with random sampling of the
-#'   distribution, users should set a seed (\code{\link{set.seed}}) for
-#'   reproducible confidence intervals. 
+#'   distribution, users should set a seed (\code{\link{set.seed}}) prior to 
+#'   calling the function for reproducible confidence intervals. 
 #'   
 #' @note 
 #'  While offsets are used to account for differences in follow-up time 
@@ -144,6 +143,16 @@
 #'   exposure. The effects do not necessarily apply across the entire range of 
 #'   the variable. However, variations in the effect are likely small, 
 #'   especially near the mean.
+#'   
+#'  @note       
+#'  Interaction terms are not allowed in the model formula. The \code{subgroup} 
+#'   argument affords interaction between the exposure variable and a single 
+#'   covariate (that is forced to categorical if supplied as numeric) to 
+#'   estimate effects of the exposure within subgroups defined by the 
+#'   interacting covariate. To include additional interaction terms with 
+#'   variables other than the exposure, we recommend that users create the 
+#'   interaction term as a cross-product of the two interaction variables in  
+#'   a data cleaning step prior to running the model.
 #'   
 #'  @note 
 #'   For negative binomial models, \code{MASS::glm.nb} is used instead of the
@@ -245,7 +254,7 @@ pointEstimate <- function(data,
   
   # Specify model family and link for the given outcome.type
   if (outcome.type %in% c("binary")) {
-    if (class(working.df[,Y]) == "factor" & length(levels(working.df[,Y])) > 2) stop("Outcome is not binary")
+    if (inherits(working.df[,Y], "factor") & length(levels(working.df[,Y])) > 2) stop("Outcome is not binary")
         family <- stats::binomial(link = 'logit')
   } else if (outcome.type %in% c("count", "rate")) {
     if (!is.numeric(working.df %>% dplyr::pull(Y))) stop("Outcome is not numeric")
@@ -320,7 +329,7 @@ pointEstimate <- function(data,
   if (!is.null(offset)) {
     working.df <- working.df %>%
       dplyr::mutate(!!offset2_name := !!offset + 0.00001) 
-    formula <- stats::as.formula(paste(paste0(as.character(formula)[2], as.character(formula)[1], as.character(formula)[3]), paste0("offset(log(", offset2_name, "))"), sep = " + "))
+    formula <- stats::as.formula(paste0(as.character(formula)[2], as.character(formula)[1], as.character(formula)[3], " + ", paste0("offset(log(", offset2_name, "))")))
     if (!outcome.type %in% c("count_nb", "rate_nb")) {
       glm_result <- stats::glm(formula = formula, data = working.df, family = family, na.action = stats::na.omit)
       # eval(parse(text = paste0(